Changes to convert `numberOfAtoms' from integer*8 to integer

+++ DOCs/TEMPLATEs/model_El_P_Template.c
@@ -115,7 +115,7 @@ static void compute(void* km, int* ier)
    int NBC;
    char* NBCstr;
 
-   intptr_t* nAtoms;
+   int* nAtoms;
    int* atomTypes;
    double* Rij_list;
    double* coords;
@@ -224,7 +224,7 @@ static void compute(void* km, int* ier)
 
 
    /* unpack data from KIM object */
-   nAtoms = (intptr_t*) KIM_API_get_data(pkim, "numberOfAtoms", ier);
+   nAtoms = (int*) KIM_API_get_data(pkim, "numberOfAtoms", ier);
    if (1 > *ier)
    {
       report_error(__LINE__, "KIM_API_get_data", *ier);
